Move main Parsley definition
