+++ src/parsley.js
@@ -4,102 +4,7 @@
 //     Parsley may be freely distributed under the MIT license.
 
 import $ from 'jquery';
-import ParsleyUtils from './parsley/utils';
-import ParsleyDefaults from './parsley/defaults';
-import ParsleyAbstract from './parsley/abstract';
-import ParsleyValidatorRegistry from './parsley/validator_registry';
-import ParsleyUI from './parsley/ui';
-import ParsleyForm from './parsley/form';
-import ParsleyField from './parsley/field';
-import ParsleyMultiple from './parsley/multiple';
-import ParsleyFactory from './parsley/factory';
+import Parsley from './parsley/main';
 import PubSub from './parsley/pubsub';
-import en from './i18n/en';
-// import i18nEN from './i18n/en';
-
-var vernums = $.fn.jquery.split('.');
-if (parseInt(vernums[0]) <= 1 && parseInt(vernums[1]) < 8) {
-  throw "The loaded version of jQuery is too old. Please upgrade to 1.8.x or better."
-}
-
-// Inherit `on`, `off` & `trigger` to Parsley:
-var Parsley = $.extend(new ParsleyAbstract(), {
-    $element: $(document),
-    actualizeOptions: null,
-    _resetOptions: null,
-    Factory: ParsleyFactory,
-    version: '@@version'
-  });
-
-// Supplement ParsleyField and Form with ParsleyAbstract
-// This way, the constructors will have access to those methods
-$.extend(ParsleyField.prototype, ParsleyAbstract.prototype);
-$.extend(ParsleyForm.prototype, ParsleyAbstract.prototype);
-// Inherit actualizeOptions and _resetOptions:
-$.extend(ParsleyFactory.prototype, ParsleyAbstract.prototype);
-
-// ### jQuery API
-// `$('.elem').parsley(options)` or `$('.elem').psly(options)`
-$.fn.parsley = $.fn.psly = function (options) {
-  if (this.length > 1) {
-    var instances = [];
-
-    this.each(function () {
-      instances.push($(this).parsley(options));
-    });
-
-    return instances;
-  }
-
-  // Return undefined if applied to non existing DOM element
-  if (!$(this).length) {
-    ParsleyUtils.warn('You must bind Parsley on an existing element.');
-
-    return;
-  }
-
-  return new ParsleyFactory(this, options);
-};
-
-// ### ParsleyField and ParsleyForm extension
-// Ensure the extension is now defined if it wasn't previously
-if ('undefined' === typeof window.ParsleyExtend)
-  window.ParsleyExtend = {};
-
-// ### Parsley config
-// Inherit from ParsleyDefault, and copy over any existing values
-Parsley.options = $.extend(ParsleyUtils.objectCreate(ParsleyDefaults), window.ParsleyConfig);
-window.ParsleyConfig = Parsley.options; // Old way of accessing global options
-
-// ### Globals
-window.Parsley = window.psly = Parsley;
-window.ParsleyUtils = ParsleyUtils;
-
-// ### Define methods that forward to the registry, and deprecate all access except through window.Parsley
-var registry = window.Parsley._validatorRegistry = new ParsleyValidatorRegistry(window.ParsleyConfig.validators, window.ParsleyConfig.i18n);
-window.ParsleyValidator = {};
-$.each('setLocale addCatalog addMessage addMessages getErrorMessage formatMessage addValidator updateValidator removeValidator'.split(' '), function (i, method) {
-  window.Parsley[method] = $.proxy(registry, method);
-  window.ParsleyValidator[method] = function () {
-    ParsleyUtils.warnOnce('Accessing the method `'+ method +'` through ParsleyValidator is deprecated. Simply call `window.Parsley.' + method + '(...)`');
-    return window.Parsley[method].apply(window.Parsley, arguments);
-  };
-});
-
-// ### ParsleyUI
-// UI is a separate class that only listens to some events and then modifies the DOM accordingly
-// Could be overriden by defining a `window.ParsleyConfig.ParsleyUI` appropriate class (with `listen()` method basically)
-window.ParsleyUI = 'function' === typeof window.ParsleyConfig.ParsleyUI ?
-  new window.ParsleyConfig.ParsleyUI().listen() : new ParsleyUI().listen();
-
-// ### PARSLEY auto-binding
-// Prevent it by setting `ParsleyConfig.autoBind` to `false`
-if (false !== window.ParsleyConfig.autoBind) {
-  $(function () {
-    // Works only on `data-parsley-validate`.
-    if ($('[data-parsley-validate]').length)
-      $('[data-parsley-validate]').parsley();
-  });
-}
 
 export default Parsley;
